requirements.txt
customtkinter
opencv-python
pandas
Pillow